+++ app/src/main/java/com/brandonjamesyoung/levelup/data/Reward.kt
@@ -1,0 +1,6 @@
+package com.brandonjamesyoung.levelup.data
+
+data class Reward(
+    val exp: Int = 0,
+    val points: Int = 0
+)